initial svelte refactor
