--- .env.template
@@ -1,1 +1,0 @@
-VITE_DEV_URL=http://localhost:8000
