--- .vim/coc-settings.json
@@ -1,4 +1,0 @@
-{
-  "volar.takeOverMode.enabled": true,
-  "tsserver.enable": false
-}
