+++ index.html
@@ -2,38 +2,9 @@
 <html lang="en">
   <head>
     <meta charset="UTF-8" />
-    <link rel="icon" type="image/webp" href="/favicon.webp" />
+    <link rel="icon" type="image/svg+xml" href="/vite.svg" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>SJVAir</title>
-    <link rel="stylesheet" href="https://fonts.googleapis.com/css2?family=Material+Symbols+Outlined:opsz,wght,FILL,GRAD@20..48,100..700,0..1,-50..200" />
-    <script>
-      // Hack to force load Material Symbols early
-      document.fonts.load('10pt "Material Symbols Outlined"');
-    </script>
-
-    <style>
-      :root {
-        --height: 325px;
-        --width: 225px;
-        --border-radius: 16px;
-        --light-blue: #1fb9ef;
-      }
-
-      * {
-        margin: 0;
-        padding: 0;
-        box-sizing: border-box;
-      }
-
-      body {
-        width: var(--width);
-        height: var(--height);
-        background-color: transparent;
-        overflow: hidden;
-        padding: 2rem;
-        font-family: Verdana, sans-serif;
-      }
-    </style>
+    <title>Vite + Svelte + TS</title>
   </head>
   <body>
     <div id="app"></div>
