+++ package.json
@@ -1,33 +1,28 @@
 {
-  "name": "@sjvair/web-widget",
-  "description": "SJVAir's monitor widget for the web!",
-  "version": "1.0.9",
+  "name": "widget-svelte",
+  "private": true,
+  "version": "0.0.0",
   "type": "module",
-  "files": [
-    "dist"
-  ],
   "scripts": {
     "dev": "vite",
-    "build": "vue-tsc && vite build",
-    "prepublishOnly": "npm run build",
+    "build": "vite build",
     "preview": "vite preview",
-    "test": "exit 0"
+    "check": "svelte-check --tsconfig ./tsconfig.app.json && tsc -p tsconfig.node.json"
   },
   "devDependencies": {
+    "@sjvair/sdk": "npm:@jsr/sjvair__sdk@^2.0.0-alpha.20",
+    "@sveltejs/vite-plugin-svelte": "^6.1.1",
+    "@tsconfig/svelte": "^5.0.4",
+    "@tstk/decorators": "npm:@jsr/tstk__decorators@^1.0.4",
+    "@tstk/utils": "npm:@jsr/tstk__utils@^1.1.1",
     "@types/gaugeJS": "npm:@types/gaugejs@^1.3.5",
     "@types/node": "^24.3.0",
-    "@vitejs/plugin-vue": "^6.0.1",
     "color2k": "^2.0.3",
     "gaugeJS": "^1.3.9",
     "sass": "^1.90.0",
+    "svelte": "^5.38.1",
+    "svelte-check": "^4.3.1",
     "typescript": "^5.9.2",
-    "vite": "^7.1.3",
-    "vue": "^3.5.20",
-    "vue-tsc": "^3.0.6"
-  },
-  "dependencies": {
-    "@sjvair/sdk": "npm:@jsr/sjvair__sdk@^2.0.0-alpha.20",
-    "@tstk/decorators": "npm:@jsr/tstk__decorators@^1.0.4",
-    "@tstk/utils": "npm:@jsr/tstk__utils@^1.1.1"
+    "vite": "^7.1.3"
   }
 }
