--- src/AirQualityGuidelines/index.ts
@@ -1,1 +1,0 @@
-export { default as AirQualityGuidelinesVue } from "./AirQualityGuidelines.vue";
