--- src/Utils/index.ts
@@ -1,2 +1,0 @@
-export * from "./asyncInitializer";
-export * from "./Colors";
