+++ src/views/Error.svelte
@@ -1,0 +1,12 @@
+<script lang="ts">
+</script>
+
+<template>
+  <h1>There was an error</h1>
+</template>
+
+<style>
+  h1 {
+    font-size: 1rem;
+  }
+</style>
