--- src/views/Error.vue
@@ -1,13 +1,0 @@
-<script setup lang="ts">
-
-</script>
-
-<template>
-  <h1>There was an error</h1>
-</template>
-
-<style scoped lang="scss">
-h1 {
-  font-size: 1rem;
-}
-</style>
