+++ vite.config.ts
@@ -1,17 +1,11 @@
-import { resolve } from 'node:path';
-import { defineConfig } from 'vite';
-import vue from '@vitejs/plugin-vue';
+import { defineConfig } from 'vite'
+import { svelte } from '@sveltejs/vite-plugin-svelte'
 
-// https://vitejs.dev/config/
+// https://vite.dev/config/
 export default defineConfig({
   base: "/static/widget",
-  plugins: [vue()],
+  plugins: [svelte()],
   esbuild: {
     target: "es2024"
   },
-  resolve: {
-    alias: {
-      '@': resolve(__dirname, './src'),
-    },
-  }
 })
